backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
django-allauth==0.63.6
dj-rest-auth==5.0.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Pillow==10.4.0
pyotp==2.9.0
qrcode==7.4.2
requests==2.32.3
cryptography==42.0.8
django-cors-headers==4.4.0
